requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==40.0.1
daphne==3.0.2
Django==3.2.18
django-cors-headers==3.14.0
django-seed==0.3.1
djangorestframework==3.13.1
Faker==18.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.5.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toposort==1.10
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
